refactoring city search field
