+++ src/components/pages/forms/search-routes-form.tsx/searchRoutesForm.tsx
@@ -2,7 +2,7 @@ import { useRouter } from "next/router";
 import { ParsedUrlQueryInput } from "querystring";
 import React, { FormEvent, useCallback } from "react";
 
-import CitySearchField from "@/components/shared/сitySearchField/сitySearchField";
+import CitySearchField from "@/components/shared/сity-search-field/сitySearchField";
 import { changeDepartureCity, changeDestinationCity } from "@/lib/features/routes/locationsSlice";
 import {
     dateEndChanged, dateStartChanged, selectDateEnd, selectDateStart, selectDefinedRoutesSettings
